+++ examples/terminal_highlight.rs
@@ -1,14 +1,13 @@
-use inkjet::{
-    formatter::{Terminal, Theme},
-    Highlighter, Language,
-};
-use std::error::Error;
 use std::fs::File;
 use std::io::{self, Read};
-use termcolor::{ColorChoice, StandardStream};
 
-mod theme_example;
-use theme_example::create_theme;
+use inkjet::{
+    formatter::Terminal,
+    theme::{Theme, vendored},
+    Highlighter, Language, InkjetError
+};
+
+use termcolor::{ColorChoice, StandardStream};
 
 fn read_file(path: &str) -> io::Result<String> {
     let mut content = String::new();
@@ -16,7 +15,7 @@ fn read_file(path: &str) -> io::Result<String> {
     Ok(content)
 }
 
-fn main() -> Result<(), Box<dyn Error>> {
+fn main() -> Result<(), InkjetError> {
     // Read the source code from a file or use a sample
     let source = read_file("example.rs").unwrap_or_else(|_| {
         r#"
@@ -33,7 +32,7 @@ fn main() {
     let lang = "rust"; // You can change this to test different languages
     let mut highlighter = Highlighter::new();
     let language = Language::from_token(lang).unwrap_or(Language::Plaintext);
-    let theme: Theme = create_theme();
+    let theme: Theme = Theme::from_helix(vendored::GITHUB_DARK)?;
 
     // Create a StandardStream for stdout
     let stream = StandardStream::stdout(ColorChoice::Always);
@@ -41,14 +40,14 @@ fn main() {
     // Create the Terminal formatter with both theme and stream
     let formatter = Terminal::new(theme, stream);
 
-    // Split the code into lines
-    let lines: Vec<&str> = source.lines().collect();
-    for line in lines {
+    for line in source.lines() {
         highlighter
             .highlight_to_writer(language, &formatter, line, &mut io::stdout())
             .map_err(|e| io::Error::new(io::ErrorKind::Other, e.to_string()))?;
         println!();
     }
+
     println!();
+    
     Ok(())
 }