+++ src/formatter/terminal.rs
@@ -2,8 +2,9 @@ use std::cell::RefCell;
 use std::io::Write;
 use termcolor::{Color, ColorSpec, StandardStream, WriteColor};
 
-use super::{Formatter, HighlightEvent, Theme};
+use super::{Formatter, HighlightEvent};
 use crate::constants::HIGHLIGHT_NAMES;
+use crate::theme::Theme;
 use crate::Result;
 
 /// A formatter for highlighting into terminal output with ANSI color codes.
@@ -23,11 +24,6 @@ impl Terminal {
             stream: RefCell::new(stream),
         }
     }
-
-    fn color_from_hex(&self, hex: &str) -> Color {
-        let rgb = color_from_hex(hex).unwrap_or((255, 255, 255));
-        Color::Rgb(rgb.0, rgb.1, rgb.2)
-    }
 }
 
 impl Formatter for Terminal {
@@ -42,8 +38,15 @@ impl Formatter for Terminal {
             }
             HighlightEvent::HighlightStart(highlight) => {
                 let style_name = HIGHLIGHT_NAMES[highlight.0];
-                let style = self.theme.get_style(style_name);
-                let color = self.color_from_hex(&style.primary_color);
+
+                let color = self
+                    .theme
+                    .get_style(style_name)
+                    .and_then(|s| s.fg)
+                    .unwrap_or(self.theme.fg);
+
+                let color = Color::Rgb(color.r, color.g, color.b);
+
                 self.stream
                     .borrow_mut()
                     .set_color(ColorSpec::new().set_fg(Some(color)))?;
@@ -62,16 +65,4 @@ impl Formatter for Terminal {
         self.stream.borrow_mut().flush()?;
         Ok(())
     }
-}
-
-fn color_from_hex(hex: &str) -> Option<(u8, u8, u8)> {
-    if hex.len() != 7 || !hex.starts_with('#') {
-        return None;
-    }
-
-    let r = u8::from_str_radix(&hex[1..3], 16).ok()?;
-    let g = u8::from_str_radix(&hex[3..5], 16).ok()?;
-    let b = u8::from_str_radix(&hex[5..7], 16).ok()?;
-
-    Some((r, g, b))
 }