+++ src/lib.rs
@@ -3,6 +3,7 @@
 //! ## Getting Started
 //! 
 //! - To get started with highlighting, take a look at the [`Highlighter`] type.
+//! - See the [`formatter`] module for the output options bundled with Inkjet, and the [`theme`] module for the optional theming API.
 //! - If you'd like to write your own formatter, see the [`Formatter`] trait.
 //! 
 //! ## Included Languages
@@ -98,7 +99,7 @@
 //! 
 //! ## Cargo Features
 //! - (Default) `html` - enables the bundled HTML formatter, which depends on `v_htmlescape` and the `theme` feature.
-//! - (Default) `theme` - enables the theme API, which depends on the `html` feature, `ahash`, `toml` and `serde`.
+//! - (Default) `theme` - enables the theme API, which depends on `ahash`, `toml` and `serde`.
 //! - (Default) `all-languages` - enables all languages.
 //! - `language-{name}` - enables the specified language.
 //!   - If you want to only enable a subset of the included languages, you'll have to set `default-features=false` and manually re-add each language you want to use.
@@ -112,6 +113,8 @@ mod error;
 mod languages;
 pub mod constants;
 pub mod formatter;
+#[cfg(feature = "theme")]
+pub mod theme;
 
 use tree_sitter_highlight::{
     Highlighter as TSHighlighter,
@@ -126,6 +129,7 @@ pub use crate::languages::Language;
 
 pub use crate::error::{
     InkjetError,
+    ThemeError,
     InkjetResult as Result
 };